requirements.txt
PyYAML
git+https://github.com/Therosin/PyTagScript